Refactor A Lot
Added in FileManager, removed previous way files were maintained as it was not that useful and was repetitive.

Refactored a lot of classes to conform to new FileManager; more refactoring needs to be done to make code look a lot cleaner.

Finished implemented arena commands, currently able to create an arena, set the player spawn and setup mob spawns.

Next functionality that should be implemented is the Game and Wave handler.

Note: Should also add in a way to delete arenas.