Fix wave functionality | Add scoreboard
Waves/Game finally work as intended. Confirmed after max waves are reached it will teleport the player to spawn and end the game. Game can be started again without any issues.

Added a scoreboard that shows default player stats. Scoreboard is dynamic and will display additional information when player is in a game.